Removed .unit because it can be inferred by the presence of the size class
The .unit class is redundant and does nothing without an additonal size class,
so why not just use the size class. I realize its kind of a long rule, but it
doesn't affect the cascade, is backward compatible with existing markup, and
saves a considerable amount of markup. It also reduces user error because it's
one less thing to remember when marking up grids.